requirements.txt
streamlit
pandas
numpy
matplotlib
rdkit-pypi
scipy
nltk